+++ src/Controller/CartController.php
@@ -9,14 +9,18 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Csrf\CsrfToken;
+use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
 
 class CartController extends AbstractController
 {
     private $entityManager;
+    private $csrfTokenManager;
 
-    public function __construct(EntityManagerInterface $entityManager)
+    public function __construct(EntityManagerInterface $entityManager, CsrfTokenManagerInterface $csrfTokenManager)
     {
         $this->entityManager = $entityManager;
+        $this->csrfTokenManager = $csrfTokenManager;
     }
 
     #[Route('/my-shopping/{reason}', name: 'app_cart', defaults: ['reason' => null])]
@@ -63,12 +67,29 @@ class CartController extends AbstractController
     }
 
     #[Route('/cart/delete/{id}', name: 'app_cart_delete')]
-    public function delete($id, Cart $cart): Response
+    public function delete($id, Cart $cart, Request $request): Response
     {
-        $cart->delete($id);
+        // // Validate CSRF token to delete a product from the list
+        // if ($this->isCsrfTokenValid('delete' . $id, $request->query->get('_token'))) {
 
-        //message
-        $this->addFlash('success', 'The product has been deleted from your shopping cart.');
+        //     $cart->delete($id);
+
+        //     //message
+        //     $this->addFlash('success', 'The product has been deleted from your shopping cart.');
+        // } else {
+        //     $this->addFlash('danger', 'You don\'t have access to it.');
+        // }
+
+        //security csrf
+        $csrfToken = new CsrfToken('delete' . $id, $request->request->get('_token'));
+        if (!$this->csrfTokenManager->isTokenValid($csrfToken)) {
+            // throw $this->createAccessDeniedException('Invalid CSRF token.');
+            $this->addFlash('danger', 'You don\'t have access to it.');
+        } else {
+            $cart->delete($id);
+            $this->addFlash('success', 'The product has been deleted from your shopping cart.');
+        }
+
 
         return $this->redirectToRoute('app_cart');
     }
